feedback from noah
