+++ packages/stateful/actions/components/TransferCw721.tsx
@@ -2,12 +2,13 @@ import {
   ArrowRightAltRounded,
   SubdirectoryArrowRightRounded,
 } from '@mui/icons-material'
-import { useEffect, useMemo, useState } from 'react'
+import { useEffect, useState } from 'react'
 import { useFormContext } from 'react-hook-form'
 import { useTranslation } from 'react-i18next'
 
 import {
   AddressInput,
+  Button,
   ImageEmoji,
   InputErrorMessage,
   NftCard,
@@ -42,24 +43,20 @@ export const TransferCw721Component: ActionComponent<TransferCw721Options> = ({
   const tokenId = watch(fieldNamePrefix + 'tokenId')
   const collection = watch(fieldNamePrefix + 'collection')
 
-  const selected = useMemo(
-    () => `${collection}:${tokenId}`,
-    [tokenId, collection]
-  )
+  const selected = `${collection}${tokenId}`
+  const getIdForNft = (nft: NftCardInfo) =>
+    `${nft.collection.address}${nft.tokenId}`
 
   useEffect(() => {
     if (!selected) {
       setError(fieldNamePrefix + 'collection', {
         type: 'required',
-        message: 'no NFT selected',
+        message: t('error.noNftSelected'),
       })
     } else {
       clearErrors(fieldNamePrefix + 'collection')
     }
-  }, [selected])
-
-  const getIdForNft = (nft: NftCardInfo) =>
-    `${nft.collection.address}:${nft.tokenId}`
+  }, [selected, setError, clearErrors, t])
 
   const [showModal, setShowModal] = useState<boolean>(isCreating)
 
@@ -73,21 +70,21 @@ export const TransferCw721Component: ActionComponent<TransferCw721Options> = ({
         <div className="flex flex-col gap-1">
           {nftInfo && <NftCard {...nftInfo} className="max-w-xs" />}
           {isCreating && (
-            <button
-              className="text-tertiary font-mono text-sm underline"
+            <Button
+              className="text-text-tertiary"
               onClick={() => setShowModal(true)}
+              variant="underline"
             >
               ({t('button.selectNft')})
-            </button>
+            </Button>
           )}
           <InputErrorMessage error={errors?.collection} />
         </div>
-        <div className="flex grow flex-row items-center gap-2 sm:gap-3">
+        <div className="flex grow flex-row items-center justify-center gap-2 sm:gap-3">
           <div className="flex flex-row items-center pl-1 sm:pl-0">
             <ArrowRightAltRounded className="!hidden !h-6 !w-6 text-text-secondary sm:!block" />
             <SubdirectoryArrowRightRounded className="!h-4 !w-4 text-text-secondary sm:!hidden" />
           </div>
-
           <div className="flex flex-col gap-1">
             <AddressInput
               containerClassName="grow"
