+++ packages/utils/nft.ts
@@ -27,3 +27,56 @@ export const normalizeNftImageUrl = (url: string) => {
 
   return url
 }
+
+// Tries to parse [EIP-721] metadata out of the data at it's metadata pointer.
+//
+// [EIP-721]: https://github.com/ethereum/EIPs/blob/master/EIPS/eip-721.md
+export const parseNftUriResponse = (
+  uriDataResponse: string,
+  collectionName: string
+): {
+  name: string | undefined
+  imageUrl: string | undefined
+  externalLink: { href: string; name: string } | undefined
+} => {
+  // Maps domain -> human readable name. If a domain is in this set, NFTs
+  // associated with it will have their external links displayed using the human
+  // readable name provided here.
+  const HostnameMap: Record<string, string | undefined> = {
+    'stargaze.zone': 'Stargaze',
+  }
+
+  let name
+  let imageUrl
+  let externalLink
+  // Only try to parse if there's a good chance this is JSON, the
+  // heuristic being the first non-whitespace character is a "{".
+  if (uriDataResponse.trimStart().startsWith('{')) {
+    try {
+      const json = JSON.parse(uriDataResponse)
+
+      if (typeof json.name === 'string' && !!json.name.trim()) {
+        name = getNftName(collectionName, json.name)
+      }
+
+      if (typeof json.image === 'string' && !!json.image) {
+        imageUrl = transformIpfsUrlToHttpsIfNecessary(json.image)
+      }
+
+      if (typeof json.external_url === 'string' && !!json.external_url.trim()) {
+        const externalUrl = transformIpfsUrlToHttpsIfNecessary(
+          json.external_url
+        )
+        const externalUrlDomain = new URL(externalUrl).hostname
+        externalLink = {
+          href: externalUrl,
+          name: HostnameMap[externalUrlDomain] ?? externalUrlDomain,
+        }
+      }
+    } catch (err) {
+      console.error(err)
+    }
+  }
+
+  return { name, imageUrl, externalLink }
+}
